Update README.md
This code defines a MainForm class that inherits from the Form class provided by Windows Forms. It sets up the UI controls (labels, text boxes, numeric up-down control, and a button) and handles the submit button click event to retrieve the entered data.When the submit button is clicked, it retrieves the name and age entered by the user and performs any required processing, such as storing the data in a database or displaying it in a message box.To use this code, you need to create a new Windows Forms Application project in Visual Studio or any other C# development environment, and then replace the default code in the Form1.cs file with the code provided above.